Separated out enemy functionality from map.h and map.c
